tests/wpt/tests/css/cssom/HTMLLinkElement-disabled-005.html

<!doctype html>
<title>&lt;link disabled&gt;'s "explicitly enabled" persists across rel changes</title>
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="author" title="Mozilla" href="https://mozilla.org">
<link rel="help" href="https://html.spec.whatwg.org/#attr-link-disabled">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link title="alt" rel="yadayada" disabled href="data:text/css,html { background: green }">
<script>
function assert_applies(applies) {
  (applies ? assert_equals : assert_not_equals)(getComputedStyle(document.documentElement).backgroundColor, "rgb(0, 128, 0)");
}

const link = document.querySelector("link[disabled]");

async_test(function(t) {
  link.onload = t.step_func_done(function() {
    assert_applies(true);
    link.setAttribute("rel", "alternate stylesheet");
    assert_applies(true);
    assert_false(link.disabled);
  });
  link.disabled = false;
  link.setAttribute("rel", "stylesheet");
}, "HTMLLinkElement.disabled's explicitly enabled state persists regardless of rel");
</script>
